:memo: :: move filter textfield component to filter feature pakage
